View update + adding retrieved data to main window

+++ src/main/java/pl/bartlomiej_swies/model/weatherData/MainConditions.java
@@ -11,8 +11,8 @@ public class MainConditions {
         this.humidity = humidity;
     }
 
-    public float getTemp() {
-        return temp;
+    public int getTemp() {
+        return (int) temp;
     }
 
     public int getPressure() {
